Make the new task form appear on the middle of the screen clearing all the previous todos and restoring them
