+++ src/style.css
@@ -127,7 +127,7 @@ transition: transform 0.4s ease;
 }
 .form, .todo-item{
   max-width: 30rem;
-  margin: 0 auto;
+  margin:  0.5em;
   padding: 1rem;
   border-radius: 8px;
   box-shadow: 0 4px 20px rgba(0, 0, 0, 0.15);
